user authorization

+++ backend/routes/bets.js
@@ -12,7 +12,8 @@ router.post(
   (req, res, next) => {
     const bet = new Bet({
       title: req.body.title,
-      content: req.body.content
+      content: req.body.content,
+      creator: req.userData.userId
     });
     bet.save().then(createdBet => {
       res.status(201).json({
@@ -33,10 +34,15 @@ router.put(
     const bet = new Bet({
       _id: req.body.id,
       title: req.body.title,
-      content: req.body.content
+      content: req.body.content,
+      userId: req.userData.userId
     });
-    Bet.updateOne({ _id: req.params.id }, bet).then(result => {
-      res.status(200).json({ message: 'update'});
+    Bet.updateOne({ _id: req.params.id, creator: req.userData.userId }, bet).then(result => {
+      if (result.nModified > 0){
+        res.status(200).json({ message: 'update'});
+      } else {
+      res.status(401).json({ message: 'Not authorized'});
+      }
     });
   }
 );
@@ -61,9 +67,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Bet.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Bet deleted!" });
+  Bet.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0){
+      res.status(200).json({ message: 'Bet deleted'});
+    } else {
+    res.status(401).json({ message: 'Not authorized'});
+    }
   });
 });
